Restructure with viewmodel
- added a viewmodel for NewGameScreen1
- moved widgets for NewGameScreen1 into separate files in a new folder
- changed NewGameScreen1 to use a viewmodel, passed to children
