+++ lib/models/view_models/new_game1.dart
@@ -1,0 +1,4 @@
+class NewGameViewModel1 {
+  String gameName;
+  String password;
+}
