+++ lib/screens/new_game1/game_name_field.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+import 'package:niira/models/view_models/new_game1.dart';
+
+class GameNameField extends StatefulWidget {
+  final NewGameViewModel1 _vm;
+  const GameNameField(this._vm, {Key key}) : super(key: key);
+
+  @override
+  _GameNameFieldState createState() => _GameNameFieldState();
+}
+
+class _GameNameFieldState extends State<GameNameField> {
+  bool _checkingName = true;
+  bool _nameAlreadyExists;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(30, 20, 30, 0),
+      child: TextFormField(
+        key: Key('name_field'),
+        decoration: InputDecoration(
+          labelText: 'Name',
+          focusedBorder: OutlineInputBorder(
+            borderSide: BorderSide(color: Colors.black, width: 2.0),
+          ),
+          enabledBorder: OutlineInputBorder(
+            borderSide: BorderSide(color: Colors.black, width: 2.0),
+          ),
+        ),
+        validator: (_) => _checkingName
+            ? 'Checking...'
+            : _nameAlreadyExists
+                ? 'Name already exists, please try another.'
+                : null,
+        onChanged: (val) => setState(() => widget._vm.gameName = val),
+      ),
+    );
+  }
+}
